# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axgit API",
    "description": "Read-only JSON API over bare Git repositories, plus Smart HTTP (fetch/clone) and an Atom feed.\n\nThere are no mutation endpoints and no authentication: writes happen over SSH on the git-server container, never here.\n\nThis document describes shapes, parameters and status codes. The semantic rules it cannot express — truncation caps, longest-ref matching for the `{ref}`/`{path}` split, merge-commit log simplification, and the exact conditions under which a field is `null` — live in `docs/API.md`, which remains the normative contract.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/repos": {
      "get": {
        "tags": [
          "repos"
        ],
        "summary": "List repositories",
        "description": "Served from the scan snapshot ([`crate::cache::ScanCache`]), not the\nresponse cache — the list has no single backing repository, so its `ETag`\nis a hash of the serialized body.",
        "operationId": "list_repos",
        "responses": {
          "200": {
            "description": "Repository list",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Hash of the response body; opaque"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReposResponse"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "500": {
            "description": "Repository root could not be scanned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/repos/{repo}": {
      "get": {
        "tags": [
          "repos"
        ],
        "summary": "Repository summary",
        "description": "An empty repository (unborn HEAD) still answers `200`, with `head`,\n`default_branch` and `last_modified` `null` and both counts 0.",
        "operationId": "get_repo",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          }
        ],
        "responses": {
          "200": {
            "description": "Repository summary",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; opaque"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoSummary"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "404": {
            "description": "`repo_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repos/{repo}/refs": {
      "get": {
        "tags": [
          "repos"
        ],
        "summary": "Branches and tags",
        "description": "Tag targets are peeled to the commit, so an annotated tag reports the\ncommit sha rather than the tag object.",
        "operationId": "get_refs",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          }
        ],
        "responses": {
          "200": {
            "description": "Branches and tags, each sorted by name",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Validator-derived; opaque"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefsInfo"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched the current `ETag`"
          },
          "404": {
            "description": "`repo_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BranchRef": {
        "type": "object",
        "description": "Branch entry of `GET /api/v1/repos/{repo}/refs` (docs/API.md).",
        "required": [
          "name",
          "target",
          "committed_at"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "main"
          },
          "target": {
            "type": "string",
            "description": "Commit sha of the branch tip."
          },
          "committed_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Authordate (RFC 3339) of the branch tip."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "One of `repo_not_found`, `ref_not_found`, `path_not_found`,\n`invalid_param`, `read_only`, `internal`.",
            "example": "repo_not_found"
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Internal errors report a generic message; the\ncause only goes to the server log.",
            "example": "repository 'foo' not found"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error envelope shared by every failing endpoint (docs/API.md).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "RefsInfo": {
        "type": "object",
        "required": [
          "branches",
          "tags"
        ],
        "properties": {
          "branches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchRef"
            },
            "description": "Sorted by name ascending; empty for a repository without commits."
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagRef"
            },
            "description": "Sorted by name ascending."
          }
        }
      },
      "RepoInfo": {
        "type": "object",
        "description": "Repository list entry as defined by `GET /api/v1/repos` in docs/API.md.",
        "required": [
          "name",
          "section",
          "owner",
          "description",
          "default_branch",
          "last_modified"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Repository name without the `.git` suffix.",
            "example": "git-compose"
          },
          "section": {
            "type": [
              "string",
              "null"
            ],
            "description": "From the repo config's `[axgit]` section, falling back to `[cgit]`."
          },
          "owner": {
            "type": [
              "string",
              "null"
            ],
            "description": "From the repo config's `[axgit]` section, falling back to `[cgit]`."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "From the repo config's `[axgit]` section, falling back to `[cgit]`."
          },
          "default_branch": {
            "type": [
              "string",
              "null"
            ],
            "description": "`None` for empty repositories (unborn HEAD)."
          },
          "last_modified": {
            "type": [
              "string",
              "null"
            ],
            "description": "Agefile (`info/web/last-modified`), falling back to the HEAD authordate.\nRFC 3339. `None` for empty repositories without an agefile.",
            "example": "2026-07-24T13:06:00+09:00"
          }
        }
      },
      "RepoSummary": {
        "type": "object",
        "description": "Repository summary as defined by `GET /api/v1/repos/{repo}` in docs/API.md.\nKept separate from [`RepoInfo`] — the list and summary responses are\ndistinct API contracts.",
        "required": [
          "name",
          "section",
          "owner",
          "description",
          "default_branch",
          "last_modified",
          "head",
          "branch_count",
          "tag_count",
          "clone_url"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "git-compose"
          },
          "section": {
            "type": [
              "string",
              "null"
            ]
          },
          "owner": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_modified": {
            "type": [
              "string",
              "null"
            ],
            "example": "2026-07-24T13:06:00+09:00"
          },
          "head": {
            "type": [
              "string",
              "null"
            ],
            "description": "HEAD commit sha. `None` for empty repositories (unborn HEAD)."
          },
          "branch_count": {
            "type": "integer",
            "minimum": 0
          },
          "tag_count": {
            "type": "integer",
            "minimum": 0
          },
          "clone_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "`{clone_url_base}/{name}.git`; `None` when no base is configured."
          }
        }
      },
      "ReposResponse": {
        "type": "object",
        "required": [
          "repos"
        ],
        "properties": {
          "repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepoInfo"
            }
          }
        }
      },
      "TagRef": {
        "type": "object",
        "description": "Tag entry of `GET /api/v1/repos/{repo}/refs` (docs/API.md).",
        "required": [
          "name",
          "target",
          "annotation",
          "tagged_at"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "v1.0.0"
          },
          "target": {
            "type": "string",
            "description": "Peeled commit sha (not the tag object), so clients can link to the commit."
          },
          "annotation": {
            "type": [
              "string",
              "null"
            ],
            "description": "First line of the tag message. `None` for lightweight tags."
          },
          "tagged_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Tagger date (RFC 3339). `None` for lightweight tags."
          }
        }
      },
...
    }
  },
...
}